+++ src/tm_remote_build/api.py
@@ -9,9 +9,10 @@ logger = logging.getLogger(__name__)
 
 
 class OpenplanetTcpSocket:
-    def __init__(self, port: int) -> None:
+    def __init__(self, port: int, host_addr: str) -> None:
         self.socket = socket(AF_INET, SOCK_STREAM)
         self.port = port
+        self.host_addr = host_addr
         self.connected = False
 
     def try_connect(self) -> bool:
@@ -20,12 +21,12 @@ class OpenplanetTcpSocket:
 
         self.socket.settimeout(0.01)
         try:
-            self.socket.connect(("localhost", self.port))
+            self.socket.connect((self.host_addr, self.port))
             logger.debug(f"Connected to {str(self.socket)}")
             self.connected = True
         except Exception as e:
             logger.debug(
-                f"Error connecting to socket on port {str(self.port)}: {str(e)}"
+                f"Error connecting to {self.host_addr}:{str(self.port)}: {str(e)}"
             )
             self.connected = False
         self.socket.settimeout(None)
@@ -76,11 +77,13 @@ class OpenplanetTcpSocket:
 
 
 class RemoteBuildAPI:
-    def __init__(self, port: int) -> None:
-        self.openplanet = OpenplanetTcpSocket(port)
-        self.data_folder = ""
+    def __init__(self, port: int, host_addr: str, data_folder: str = "") -> None:
+        self.openplanet = OpenplanetTcpSocket(port, host_addr)
+        self.data_folder = data_folder
         self.op_log = OpenplanetLog()
-        self.get_data_folder()
+        if not self.data_folder:
+            self.get_data_folder()
+        self.op_log.set_path(os.path.join(self.data_folder, "Openplanet.log"))
 
     def send_route(self, route: str, data: dict) -> dict:
         response = {}
@@ -101,12 +104,8 @@ class RemoteBuildAPI:
     def get_data_folder(self) -> bool:
         if not self.get_status():
             return False
-
         response = self.send_route("get_data_folder", {})
-        response_data_folder = response.get("data", "")
-        if os.path.isdir(response_data_folder):
-            self.data_folder = response_data_folder
-            self.op_log.set_path(os.path.join(self.data_folder, "Openplanet.log"))
+        self.data_folder = response.get("data", "")
         return self.data_folder != ""
 
     def load_plugin(